+++ HISTORY.rst
@@ -5,6 +5,8 @@ History
 ##### (##########)
 ------------------
 
+* Fix an unfixed random state in one of the classification tests
+
 0.6.5 (2023-06-06)
 ------------------